move LoadedItem into its own File
