+++ Assets/Scripts/saving/LoadedItem.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c472e8633b93d6feb8007a2edcd3de76
