+++ Assets/Scripts/saving/Loader.cs
@@ -140,80 +140,6 @@ class Loader
 
 }
 
-class LoadedItem
-{
-
-    Dictionary<string,LoadedItem> dict;
-    List<LoadedItem> list;
-
-    string rawData;
-    public LoadedItem(Dictionary<string, LoadedItem> data)
-    {
-        dict=data;
-    }
-
-    public LoadedItem(List<LoadedItem> data)
-    {
-        list=data;
-    }
-
-    public LoadedItem(string data)
-    {
-        rawData=data;
-    }
-
-    bool isDict()
-    {
-        return dict!=null;
-    }
-
-    bool isList()
-    {
-        return list!=null;
-    }
-
-    bool isValue()
-    {
-        return rawData!=null;
-    }
-
-    string dataToString()
-    {
-        if(isDict())
-            return dict.ToString();
-        else if(isList())
-            return list.ToString();
-        else
-            return rawData;
-    }
-
-    public LoadedItem this[int key]
-    {
-        get
-        {
-            if(!isList())
-                throw new LoadException($"{dataToString()} is not a List");
-            return list[key];
-        }
-    }
-
-
-    public LoadedItem this[string key]
-    {
-        get
-        {
-            if(!isDict())
-                throw new LoadException($"{dataToString()} is not a Dictionary");
-            return dict[key];
-        }
-    }
-
-    T convertTo<T>()
-    {
-        throw new LoadException($"cannot convert {dataToString()} to type {typeof(T).Name}");
-    }
-}
-
 class LoadException : Exception
 {
     public LoadException()
